feat: add build a flat package
This tries to solve the problem mentioned in this
[issue](#353)

It was removed the `/dist` folder from the entry points in the `package.json `
file, instead, the folder will be set in the `rollupjs ` context, cause now
when the build process occurs, the `package.json ` file will be copied to the
dist folder.

It also adds a dist command that runs an `npm publish`  inside the `dist` folder.

> also adds the .tool-versions in .gitignore file for a better asdf usage.